backend/requirements.txt
Django==4.2.16
djangorestframework==3.14.0
channels==4.0.0
psycopg2-binary==2.9.7
redis==4.6.0
celery==5.3.1
Pillow==10.0.0
python-decouple==3.8
django-storages==1.14.1
boto3==1.28.25
openai==1.3.0
requests==2.31.0
pytest==7.4.0
pytest-django==4.5.2
factory-boy==3.3.0
django-cors-headers==4.2.0
channels-redis==4.1.0
stripe==5.5.0
gunicorn==21.2.0
whitenoise==6.5.0
coverage==7.4.1
numpy==1.24.3
